models/url_mapping.py
# TODO: LUIS - Database Model for URL Mappings
"""
TASKS FOR LUIS:
1. Define URL mapping data structure
2. Create functions for database operations (CRUD)
3. Set up Firestore document structure
4. Add validation methods
5. Create helper functions for teammates

REQUIRED FUNCTIONS FOR TEAM INTEGRATION:
- create_url_mapping(original_url, short_code, metadata)
- get_url_mapping(short_code) -> returns original_url or None
- increment_click_count(short_code)
- validate_short_code_exists(short_code) -> boolean
- get_url_stats(short_code) -> returns stats dict

FIRESTORE DOCUMENT STRUCTURE:
{
    "short_code": "abc123",
    "original_url": "https://example.com/long/url",
    "created_at": "2025-08-04T10:30:00Z",
    "click_count": 0,
    "is_active": true,
    "created_by_ip": "192.168.1.1",
    "expires_at": null  # optional expiration
}

COLLECTION NAME: "url_mappings"

EXAMPLE FUNCTIONS STRUCTURE:
from google.cloud import firestore
from datetime import datetime

class URLMapping:
    def __init__(self):
        # Initialize Firestore client
        pass
    
    def create_mapping(self, original_url, short_code, client_ip=None):
        # Create new URL mapping in Firestore
        pass
    
    def get_mapping(self, short_code):
        # Retrieve URL mapping by short code
        # Return dict with original_url or None if not found
        pass
    
    def increment_clicks(self, short_code):
        # Increment click counter for analytics
        pass
"""

# LUIS: Replace this comment block with actual URL mapping model code
